pyproject.toml
[build-system]
requires = ["setuptools>=71.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "contrastlens"
version = "0.1.0"
description = "Stochastic luminance-based image contrast renderer with a Tkinter GUI."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["image-processing", "contrast", "stochastic", "halftone", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "Pillow>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
contrastlens = "contrastlens.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["contrastlens*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["contrastlens"]
omit = ["tests/*", "main.py"]

[tool.coverage.report]
show_missing = true
